app/requirements.txt
mysql-connector-python==9.1.0
uvicorn
fastapi